Task dataclass now uses float instead of int for timestamp representation to increase the entropy

+++ sergeant/objects.py
@@ -8,7 +8,7 @@ class Task:
     kwargs: typing.Dict[str, typing.Any] = dataclasses.field(
         default_factory=dict,
     )
-    date: int = dataclasses.field(
-        default_factory=lambda: int(time.time()),
+    date: float = dataclasses.field(
+        default_factory=lambda: time.time(),
     )
     run_count: int = 0
